Reorganize dmpr into multiple files
looks better, shorter scroll paths ^^
