+++ dmpr/dmpr.py
@@ -1,88 +1,20 @@
-import ipaddress
-import random
 import copy
 import functools
-
+import ipaddress
+import random
 from datetime import datetime
 
-# example configuration for DMPR daemon
-exa_conf = """
-    "id" : "ace80ef4-d284-11e6-bf26-cec0c932ce01",
-    "rtn-msg-interval" : "30",
-    "rtn-msg-interval-jitter" : "7",
-    "rtn-msg-hold-time" : "90",
-    "max-full-update-interval": "10"
-    "mcast-v4-tx-addr" : "224.0.1.1",
-    "mcast-v6-tx-addr" : "ff05:0:0:0:0:0:0:2",
-    "proto-transport-enable"  : [ "v4" ],
-    "interfaces" : [
-      { "name" : "wlan0", "addr-v4" : "10.0.0.1", "link-characteristics" : { "bandwidth" : "100000", "loss" : "0"  } },
-      { "name" : "tetra0", "addr-v4" : "10.0.0.1", "link-characteristics" : { "bandwidth" : "10000",  "loss" : "0"  } }
-    ],
-    "networks" : [
-       { "proto": "v4", "prefix" : "192.168.1.0", "prefix-len" : "24" },
-       { "proto": "v4", "prefix" : "192.168.2.0", "prefix-len" : "24" },
-       { "proto": "v4", "prefix" : "10.10.0.0",   "prefix-len" : "16" },
-       { "proto": "v6", "prefix" : "fdcb:523:1111::", "prefix-len" : "48" },
-       { "proto": "v6", "prefix" : "fd6a:6ad:b07f:ffff::", "prefix-len" : "64" }
-    }
-"""
+from .config import DefaultConfiguration
+from .exceptions import ConfigurationException, InvalidPartialUpdate, \
+    InvalidMessage
+from .message import Message
+from .path import Path
+from .policies import AbstractPolicy
 
 
 @functools.lru_cache(maxsize=1024)
 def normalize_network(network):
     return str(ipaddress.ip_network(network, strict=False))
-
-
-def dict_reverse_lookup(d: dict, value):
-    return list(d.keys())[list(d.values()).index(value)]
-
-
-class ConfigurationException(Exception):
-    pass
-
-
-class InternalException(Exception):
-    pass
-
-
-class DMPRConfigDefaults(object):
-    rtn_msg_interval = 30
-    rtn_msg_interval_jitter = rtn_msg_interval / 4
-    rtn_msg_hold_time = rtn_msg_interval * 3
-    retracted_prefix_hold_time = rtn_msg_interval * 12  # TODO TBD
-    max_full_update_interval = 0  # 0 => disables partial updates
-
-    DEFAULT_CONFIG = {
-        'rtn-msg-interval': rtn_msg_hold_time,
-        'rtn-msg-interval-jitter': rtn_msg_interval_jitter,
-        'rtn-msg-hold-time': rtn_msg_hold_time,
-        'retracted-prefix-hold-time': retracted_prefix_hold_time,
-        'max-full-update-interval': max_full_update_interval,
-    }
-
-    # default bandwidth for a given interface in bytes/second
-    # bytes/second enabled dmpr deployed in low bandwidth environments
-    # normally this value should be fetched from a interface information
-    # or by active measurements.
-    # Implementers SHOULD quantise values into a few classes to reduce the
-    # DMPR routing packet size.
-    # E.g. 1000, 5000, 10000, 100000, 1000000, 100000000, 100000000
-    LINK_CHARACTERISITCS_BANDWIDTH = 5000
-    # default loss is in percent for a given path
-    # Implementers SHOULD quantise values into a few classes to reduce the
-    # DMPR routing packet size.
-    # e.g. 0, 5, 10, 20, 40, 80
-    LINK_CHARACTERISITCS_LOSS = 0
-    # default link cost in a hypothetical currency, the higher the more valuable
-    # e.g. wifi can be 0, LTE can be 100, satelite uplink can be 1000
-    LINK_CHARACTERISITCS_COST = 0
-
-    DEFAULT_ATTRIBUTES = {
-        'bandwidth': LINK_CHARACTERISITCS_BANDWIDTH,
-        'loss': LINK_CHARACTERISITCS_LOSS,
-        'cost': LINK_CHARACTERISITCS_COST,
-    }
 
 
 class NoOpTracer:
@@ -107,354 +39,6 @@ class NoOpLogger:
 
     def log(self, msg, sev, time=lambda: datetime.now().isoformat()):
         pass
-
-
-class Path:
-    """Auxiliary class to handle paths, appending and their link
-        attributes """
-
-    def __init__(self, path: str, attributes: dict,
-                 next_hop: str, next_hop_interface: str):
-        path = path.split('>')
-        if len(path) % 2 != 1:
-            raise InternalException("Invalid path: {}".format(path))
-
-        self.links = []
-        self.nodes = []
-        for i, element in enumerate(path):
-            if i % 2 == 0:
-                # This is a node
-                self.nodes.append(element)
-            else:
-                # This is a link
-                if not (element.startswith('[') and element.endswith(']')):
-                    msg = "Invalid path: {}, link format error: {}"
-                    raise InternalException(msg.format(path, element))
-                self.links.append(element.strip('[]'))
-
-        self.attributes = attributes
-        self.next_hop = next_hop
-        self.next_hop_interface = next_hop_interface
-        self._global_attributes = {}
-        self.policy_cache = {}
-
-    @staticmethod
-    def _next_id(attributes: dict) -> str:
-        return str(int(max(attributes, key=int, default=0)) + 1)
-
-    def append(self, node: str, new_next_hop_interface: str, attributes: dict):
-        """Append a node with link attributes to the front of the path"""
-        self.next_hop_interface = new_next_hop_interface
-        self.next_hop = self.nodes[0]
-
-        attribute_id = self._next_id(self.attributes)
-        self.links.insert(0, attribute_id)
-        self.nodes.insert(0, node)
-        self.attributes[attribute_id] = attributes
-        self.policy_cache = {}
-
-    def apply_attributes(self, attributes: dict):
-        """append all necessary attributes to global attributes dictionary"""
-        for link in self.links:
-            attr = self.attributes[link]
-            if attr not in attributes.values():
-                attribute_id = self._next_id(attributes)
-                attributes[attribute_id] = attr
-
-        self._global_attributes = attributes
-
-    def __str__(self) -> str:
-        """Returns path as string, requires an up-to-date
-            global attribute dictionary"""
-        nodes = self.nodes[:]
-        links = self.links[:]
-        result = nodes.pop()
-        while nodes:
-            next_link = links.pop()
-            next_node = nodes.pop()
-
-            attr = self.attributes[next_link]
-            try:
-                link_id = dict_reverse_lookup(self._global_attributes, attr)
-            except IndexError:
-                raise InternalException("path.__str__ called before all"
-                                        "attributes were applied")
-
-            result = "{}>[{}]>{}".format(next_node, link_id, result)
-
-        return result
-
-    def __eq__(self, other):
-        if not isinstance(other, Path):
-            raise ValueError("cannot compare path to {}".format(type(other)))
-
-        if self.nodes != other.nodes:
-            return False
-        if (self.next_hop != other.next_hop or
-                    self.next_hop_interface != other.next_hop_interface):
-            return False
-
-        for self_link, other_link in zip(self.links, other.links):
-            if self.attributes[self_link] != other.attributes[other_link]:
-                return False
-
-        return True
-
-
-class BaseMessage:
-    def __init__(self):
-        self.id = None
-        self.seq = float('-inf')
-        self.addr_v4 = None
-        self.addr_v6 = None
-
-        self.networks = {}
-        self.routing_data = {}
-        self.node_data = {}
-
-    def apply_base_data(self, msg):
-        self.id = msg.id
-        self.seq = msg.seq
-        self.addr_v4 = msg.addr_v4
-        self.addr_v6 = msg.addr_v6
-
-        self.networks = msg.networks
-        self.routing_data = msg.routing_data
-        self.node_data = msg.node_data
-
-
-class InvalidMessage(Exception):
-    pass
-
-
-class InvalidPartialUpdate(Exception):
-    pass
-
-
-class Message(BaseMessage):
-    def __init__(self, msg: dict, interface: dict, router_id: str, rx_time):
-        super(Message, self).__init__()
-        self._base = BaseMessage()
-
-        self.rx_time = rx_time
-        self.router_id = router_id
-        self.interface = interface
-        self.apply_new_msg(msg, rx_time)
-
-    def apply_new_msg(self, msg: dict, rx_time) -> bool:
-        self._validate_msg(msg)
-
-        self.rx_time = rx_time
-
-        if msg['type'] == 'full':
-            return self._apply_full(msg)
-        elif msg['type'] == 'partial':
-            return self._apply_partial(msg)
-
-    def _validate_msg(self, msg: dict):
-        # Check for essential fields
-        for i in ('id', 'seq', 'type'):
-            if i not in msg:
-                raise InvalidMessage("a message needs field {}".format(i))
-
-        if msg['id'] == self.router_id:
-            raise InvalidMessage("Message from self")
-
-        if msg['seq'] <= self.seq:
-            raise InvalidMessage("Old sequence number")
-
-        if msg['type'] == 'full':
-            # Check for at least one address
-            if not any(('addr-v4' in msg, 'addr-v6' in msg)):
-                raise InvalidMessage("Need one of add addr-v4 and addr-v6 in "
-                                     "message")
-
-        # Partial messages need a base
-        elif msg['type'] == 'partial':
-            if 'partial-base' not in msg:
-                raise InvalidMessage("Need a partial-base in partial updates")
-            if msg['partial-base'] != self._base.seq:
-                raise InvalidPartialUpdate("Wrong base message sequence number")
-
-        # If routing-data exists and there is at least one policy with paths
-        # then there must be link-attributes
-        # Also every policy value must be a dict
-        routing_data = msg.get('routing-data', {})
-        if routing_data and any(routing_data.values()):
-            if not msg.get('link-attributes'):
-                raise InvalidMessage("Need link-attributes for routing-data")
-            for i in routing_data.values():
-                if not isinstance(i, dict):
-                    raise InvalidMessage("routing-data policies must be dict")
-
-        for i in ('routing-data', 'node-data', 'link-attributes'):
-            if i in msg:
-                if not isinstance(msg[i], dict):
-                    raise InvalidMessage("{} must be dict".format(i))
-
-    def _apply_full(self, msg: dict) -> bool:
-        update = False
-        self.seq = msg['seq']
-        self.addr_v4 = msg.get('addr-v4', None)
-        self.addr_v6 = msg.get('addr-v6', None)
-
-        update |= self._save_networks(msg)
-
-        msg_routing_data = msg.get('routing-data', {})
-        msg_link_attributes = msg.get('link-attributes', {})
-        msg_node_data = msg.get('node-data', {})
-        new_routing_data = {}
-        for policy in msg_routing_data:
-            new_routing_data[policy] = {}
-            for node in msg_routing_data[policy]:
-                path = self._get_path(msg_routing_data[policy][node]['path'],
-                                      msg_link_attributes)
-                if not path:
-                    # Loop detection
-                    continue
-
-                new_routing_data[policy][node] = {
-                    'path': path,
-                }
-
-                if node not in msg_node_data:
-                    # We have no node_data for this node, apparently it does not
-                    # advertise any networks
-                    continue
-
-        if self.routing_data != new_routing_data:
-            update = True
-            self.routing_data = new_routing_data
-
-        if self.node_data != msg_node_data:
-            update = True
-            self.node_data = msg_node_data
-
-        self._base = BaseMessage()
-        self._base.apply_base_data(self)
-
-        return update
-
-    def _apply_partial(self, msg: dict) -> bool:
-        self.seq = msg['seq']
-        self.addr_v4 = msg.get('addr_v4', self._base.addr_v4)
-        self.addr_v6 = msg.get('addr_v6', self._base.addr_v6)
-
-        if 'networks' in msg:
-            self.networks = msg.get('networks', {})
-        else:
-            self.networks = self._base.networks
-
-        self.routing_data = copy.deepcopy(self._base.routing_data)
-        routing_data = msg.get('routing-data', {})
-        for policy in routing_data:
-            for node in routing_data[policy]:
-                if routing_data[policy][node] is None:
-                    if node in self.routing_data[policy]:
-                        del self.routing_data[policy][node]
-                else:
-                    path = self._get_path(routing_data[policy][node]['path'],
-                                          msg.get('link-attributes', {}))
-                    if not path:
-                        # Routing loop
-                        continue
-                    self.routing_data[policy][node] = {
-                        'path': path
-                    }
-
-        self.node_data = copy.deepcopy(self._base.node_data)
-        node_data = msg.get('node-data', {})
-        for node in node_data:
-            if node == self.router_id:
-                continue
-            if node_data[node] is None:
-                if node in self.node_data:
-                    del self.node_data[node]
-            else:
-                self.node_data[node] = node_data[node]
-
-        return True  # TODO add support for update detection
-
-    def _save_networks(self, msg: dict) -> bool:
-        networks = msg.get('networks')
-        if networks != self.networks:
-            self.networks = networks
-            return True
-        return False
-
-    def _get_path(self, path: str, link_attributes: dict) -> Path:
-        path = Path(path=path,
-                    attributes=link_attributes,
-                    next_hop='',
-                    next_hop_interface=self.interface['name'])
-
-        if self.router_id in path.nodes:
-            return False
-
-        path.append(self.router_id,
-                    self.interface['name'],
-                    self.interface['link-attributes'])
-
-        return path
-
-
-class AbstractPolicy:
-    """ A policy implements all routing decisions by
-        providing comparison methods for interfaces
-        and paths"""
-    name = NotImplemented
-
-    def __init__(self):
-        if self.name == NotImplemented:
-            raise NotImplementedError("The policy needs a name")
-
-    @staticmethod
-    def with_path_cache(func):
-        @functools.wraps(func)
-        def wrapper(self, path):
-            if self.name in path.policy_cache:
-                return path.policy_cache[self.name]
-
-            metric = func(self, path)
-            path.policy_cache[self.name] = metric
-            return metric
-
-        return wrapper
-
-    def path_cmp_key(self, path: Path):
-        """ returns a sort key e.g. for the `sorted` function"""
-        raise NotImplementedError("A policy must override this method")
-
-
-class SimpleLossPolicy(AbstractPolicy):
-    """ Route via the lowest-loss path"""
-    name = 'lowest-loss'
-
-    @staticmethod
-    def _acc_loss(path: Path) -> int:
-        loss = 0
-        for link in path.links:
-            loss += path.attributes[link]['loss']
-        return loss
-
-    @AbstractPolicy.with_path_cache
-    def path_cmp_key(self, path: Path) -> int:
-        return self._acc_loss(path)
-
-
-class SimpleBandwidthPolicy(AbstractPolicy):
-    """ Route via the highest-bandwidth path"""
-    name = 'highest-bandwidth'
-
-    @staticmethod
-    def _acc_bw(path: Path) -> int:
-        return min(path.attributes[link]['bandwidth'] for link in path.links)
-
-    @AbstractPolicy.with_path_cache
-    def path_cmp_key(self, path: Path):
-        # the minimum bandwidth of the path while slightly preferring
-        # shorter paths
-        return - self._acc_bw(path) * 0.99 ** len(path.links)
 
 
 class DMPRState:
@@ -555,7 +139,7 @@ class DMPR(object):
         if not isinstance(configuration, dict):
             raise ConfigurationException("configuration must be dict-like")
 
-        config = copy.deepcopy(DMPRConfigDefaults.DEFAULT_CONFIG)
+        config = copy.deepcopy(DefaultConfiguration.DEFAULT_CONFIG)
 
         config.update(configuration)
 
@@ -591,7 +175,7 @@ class DMPR(object):
             converted_interfaces[interface_data['name']] = interface_data
 
             orig_attr = interface_data.setdefault('link-attributes', {})
-            attributes = copy.deepcopy(DMPRConfigDefaults.DEFAULT_ATTRIBUTES)
+            attributes = copy.deepcopy(DefaultConfiguration.DEFAULT_ATTRIBUTES)
             attributes.update(orig_attr)
             interface_data['link-attributes'] = attributes
