+++ dmpr/exceptions.py
@@ -1,0 +1,14 @@
+class ConfigurationException(Exception):
+    pass
+
+
+class InternalException(Exception):
+    pass
+
+
+class InvalidMessage(Exception):
+    pass
+
+
+class InvalidPartialUpdate(Exception):
+    pass
